+++ src/MainPage/MusicPlayer/MusicPlayer.jsx
@@ -1,11 +1,11 @@
 import classes from './MusicPlayer.module.css';
-import ProgressBar from './ProgressBar/ProgressBar';
-import PlayerInterface from './PlayerInterface/PlayerInterface';
+import MusicPlayer__ProgressBar from '../MusicPlayer__ProgressBar/MusicPlayer__ProgressBar';
+import MusicPlayer__Interface from '../MusicPlayer__Interface/MusicPlayer__Interface';
 const MusicPlayer = () => {
   return (
     <div className={classes.musicPlayer}>
-      <ProgressBar />
-      <PlayerInterface />
+      <MusicPlayer__ProgressBar />
+      <MusicPlayer__Interface />
     </div>
   );
 };
