--- src/MainPage/MusicPlayer/ProgressBar/ProgressBar.jsx
@@ -1,9 +1,0 @@
-import classes from './ProgressBar.module.css';
-
-const ProgressBar = () => {
-  return (
-<div className={classes.progressBar}></div>
-  );
-};
-
-export default ProgressBar;
