+++ src/MainPage/MusicPlayer__ProgressBar/MusicPlayer__ProgressBar.jsx
@@ -1,0 +1,9 @@
+import classes from './MusicPlayer__ProgressBar.module.css';
+
+const MusicPlayer__ProgressBar = () => {
+  return (
+<div className={classes.progressBar}></div>
+  );
+};
+
+export default MusicPlayer__ProgressBar;
