+++ src/MainPage/MusicPlayer__ProgressBar/MusicPlayer__ProgressBar.module.css
