+++ src/MainPage/MusicPlayer__Volume/MusicPlayer__Volume.module.css
